Split :common:ui:resources up
This splits up the module and moves anything related to strings
into the strings module.
