--- common/ui/resources/build.gradle.kts
@@ -1,44 +1,0 @@
-// Copyright 2023, Google LLC, Christopher Banes and the Tivi project contributors
-// SPDX-License-Identifier: Apache-2.0
-
-
-plugins {
-    id("app.tivi.android.library")
-    id("app.tivi.kotlin.multiplatform")
-    alias(libs.plugins.composeMultiplatform)
-}
-
-kotlin {
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                api(projects.data.models)
-                api(projects.common.ui.strings)
-                api(compose.ui)
-            }
-        }
-
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-                implementation(libs.assertk)
-                implementation(libs.lyricist.library)
-            }
-        }
-
-        val iosMain by getting {
-            dependencies {
-                @OptIn(org.jetbrains.compose.ExperimentalComposeLibrary::class)
-                implementation(compose.components.resources)
-            }
-        }
-    }
-}
-
-android {
-    namespace = "app.tivi.common.ui.resources"
-
-    sourceSets["main"].apply {
-        res.srcDirs("src/androidMain/res", "src/commonMain/resources")
-    }
-}
