+++ common/ui/resources/fonts/src/commonMain/resources/font/inter_regular.ttf
