+++ common/ui/resources/strings/src/jvmMain/kotlin/app/tivi/common/ui/resources/String.kt
@@ -1,0 +1,6 @@
+// Copyright 2023, Christopher Banes and the Tivi project contributors
+// SPDX-License-Identifier: Apache-2.0
+
+package app.tivi.common.ui.resources
+
+actual fun String.fmt(vararg args: Any?): String = format(*args)
