--- common/ui/strings/build.gradle.kts
@@ -1,17 +1,0 @@
-// Copyright 2023, Google LLC, Christopher Banes and the Tivi project contributors
-// SPDX-License-Identifier: Apache-2.0
-
-
-plugins {
-    id("app.tivi.kotlin.multiplatform")
-}
-
-kotlin {
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                api(libs.lyricist.core)
-            }
-        }
-    }
-}
