+++ settings.gradle.kts
@@ -81,8 +81,8 @@ include(
     ":core:powercontroller",
     ":core:preferences",
     ":common:ui:circuit-overlay",
-    ":common:ui:resources",
-    ":common:ui:strings",
+    ":common:ui:resources:fonts",
+    ":common:ui:resources:strings",
     ":common:ui:compose",
     ":common:ui:screens",
     ":common:imageloading",
